fix(core): correct PlanParser chunk-splitting drift
PlanParser.fromMarkdown treated every `##` heading as a chunk, so prose
sections ("Risks", "Architecture overview", …) were parsed as chunks, and
"Chunk 3.5" truncated to id 3 and collided with the real Chunk 3 — the live
PLAN.md split into 23 chunks with duplicate ids.

- Explicit-chunk mode: when a plan uses `Chunk N` headings, only those start a
  chunk; non-chunk prose sections are skipped. Plans with no `Chunk N` headings
  keep the prior sequential-id behaviour.
- Decimal chunk ids: chunk_id/depends_on schemas relaxed to z.number().min(1);
  id, title, and dependency parsing handle fractional numbers (so a chunk can
  be inserted between two existing ones without renumbering).

The live PLAN.md now parses to 17 unique ids [1-9, 3.5, 10-16].

Note: the file-split machinery (advance() active_chunk+1, PLAN_N.md filename
pattern) still assumes integer ids; decimal split-to-file is out of scope here.

+++ packages/core/src/schemas/plan-chunk.ts
@@ -9,12 +9,16 @@ import { TaskSchema } from "./task.js";
  */
 export const PlanChunkSchema = z
 	.object({
-		/** The numeric chunk ID (matches the N in PLAN_N.md). */
-		chunk_id: z.number().int().min(1),
+		/**
+		 * The numeric chunk ID. Usually the integer N in PLAN_N.md, but fractional
+		 * IDs (e.g. `3.5`) are allowed so an interstitial chunk can be inserted
+		 * between two existing chunks without renumbering the whole plan.
+		 */
+		chunk_id: z.number().min(1),
 		/** The human-readable title of this chunk (e.g., "Security utilities"). */
 		title: z.string().min(1),
-		/** Chunk IDs that must be completed before this one can start. */
-		depends_on: z.array(z.number().int().min(1)).default([]),
+		/** Chunk IDs that must be completed before this one can start (may be fractional). */
+		depends_on: z.array(z.number().min(1)).default([]),
 		/** Estimated number of sessions to complete this chunk. */
 		est_sessions: z.number().int().min(1).optional(),
 		/** The list of tasks within this chunk. */
